requirements.txt
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
importlib-metadata==4.0.1
mysql-connector-python==8.0.24
protobuf==3.15.8
pydantic==1.8.1
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.4.11
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
zipp==3.4.1
